requirements.txt
tensorboard >= 2.18.0
torch >= 2.5.1
torchaudio >= 2.5.1
torchvision >= 0.20.1
tqdm >= 4.67.1
pyyaml >= 6.0.2